% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickets.R
\name{get_all_tickets}
\alias{get_all_tickets}
\title{Download tickets}
\usage{
get_all_tickets(start_time = 0)
}
\arguments{
\item{start_time}{An integer Unix-time representation. This limits query
results to tickets modified after that time. Default is \code{0}, meaning return
all.}
}
\value{
A list of tickets.
}
\description{
Download tickets
}
\seealso{
This uses the "incremental export" API (\url{https://developer.zendesk.com/rest_api/docs/core/incremental_export#incremental-ticket-export}).
}
