% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users}
\alias{get_users}
\title{Get all Zendesk users}
\usage{
get_users()
}
\value{
A list of users
}
\description{
Get all Zendesk users
}
\seealso{
\url{https://developer.zendesk.com/rest_api/docs/core/users}
}
