% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zd_api}
\alias{zd_api}
\alias{zd_GET}
\alias{zd_PUT}
\alias{zd_PATCH}
\alias{zd_POST}
\alias{zd_DELETE}
\title{HTTP methods for the Zendesk API}
\usage{
zd_api(verb, url, config = list(), ...)

zd_GET(url, ...)

zd_PUT(url, ...)

zd_PATCH(url, ...)

zd_POST(url, ...)

zd_DELETE(url, ...)
}
\description{
HTTP methods for the Zendesk API
}
