% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R, R/tickets.R
\name{zd_search}
\alias{zd_search}
\alias{search_tickets}
\title{Search your Zendesk site}
\usage{
zd_search(query)

search_tickets(query)
}
\arguments{
\item{query}{character search string, as you'd type in the web app}
}
\value{
A list of search results.
}
\description{
\code{zd_search()} is a general search function. \code{search_tickets()} restricts
the query to just tickets.
}
