% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zd_api}
\alias{zd_api}
\alias{zd_GET}
\alias{zd_PUT}
\alias{zd_PATCH}
\alias{zd_POST}
\alias{zd_DELETE}
\title{HTTP methods for the Zendesk API}
\usage{
zd_api(verb, url, config = list(), ...)

zd_GET(url, ...)

zd_PUT(url, ...)

zd_PATCH(url, ...)

zd_POST(url, ...)

zd_DELETE(url, ...)
}
\arguments{
\item{verb}{for \code{zd_api()}, the string HTTP method to call}

\item{url}{string URL to request}

\item{config}{list or \code{httr::config()} object with any custom request
headers, for example.}

\item{...}{Additional arguments to pass to the \code{httr} functions, such as
\code{query} or \code{body}}
}
\value{
The parsed response body content, if the request was successful.
}
\description{
\code{zd_GET()}, \code{zd_POST()}, and the other HTTP method functions are convenience
wrappers around the \code{httr} functions. They include configuration for this
API, and they understand how to handle the responses that the API returns.
\code{zd_api()} is the function they call internally.
}
